refactor(vms/evm/sync): make network.Client a generic Dispatcher
- Replaces the bundled `network.Client` with a generic
  `Dispatcher[Req, Resp proto.Message]` and per-RPC type aliases.
- External consumers can register their own RPC types by instantiating
  Dispatcher against a new handler ID without forking this package.
